requirements.txt
tweepy==4.14.0
requests==2.31.0
schedule==1.2.1
python-dotenv==1.0.0
Pillow==10.1.0
torch==2.2.0
transformers==4.36.2
openai==0.28.0
